requirements.txt
django-ckeditor


psycopg2-binary

django
django-bootstrap5

Pillow

